+++ src/utils/index.ts
@@ -11,3 +11,4 @@ export * from './ui';
 export * from './banner';
 export * from './scripts';
 export * from './module-loader';
+export * from './repo-alias';